Push getters to separate files
