+++ src/store/getActiveToolsForElement.js
@@ -1,0 +1,5 @@
+export default function (element, tools) {
+  return tools.filter(
+    (tool) => tool.element === element && tool.mode === 'active'
+  );
+}
