+++ src/store/getInteractiveToolsForElement.js
@@ -1,0 +1,7 @@
+export default function (element, tools) {
+  return tools.filter(
+    (tool) =>
+      tool.element === element &&
+      (tool.mode === 'active' || tool.mode === 'passive')
+  );
+}
