+++ src/store/getToolsWithDataForElement.js
@@ -1,0 +1,9 @@
+import { getToolState } from './../stateManagement/toolState.js';
+
+export default function (element, tools) {
+  return tools.filter((tool) => {
+    const toolState = getToolState(element, tool.name);
+
+    return toolState && toolState.data.length > 0;
+  });
+}
